missiong data type

+++ pkg/model/employee_model.go
@@ -6,7 +6,7 @@ type Employee struct {
 	ID        string
 	FirstName string
 	LastName  string
-	Salary    int64
+	Salary    float64
 	CommissionPct float32
 	HireDate  time.Time
 }
